fix: 🐛 rm dead save code
